--- client/app/2.0/utils/index.js
@@ -1,4 +1,0 @@
-/* eslint-disable  */
-export * from 'utils/reader/index';
-
-/* eslint-enable  */
