+++ client/app/reader/store/appeal.js
@@ -1,10 +1,9 @@
 // External Dependencies
-import { createSlice, createAsyncThunk, current } from '@reduxjs/toolkit';
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 
 // Local Dependencies
-import { addMetaLabel } from 'utils/reader';
-import { ENDPOINT_NAMES } from 'store/constants/reader';
 import ApiUtil from 'app/util/ApiUtil';
+import { ENDPOINT_NAMES } from './constants/reader';
 
 /**
  * Case Select Initial State
