+++ client/app/reader/store/selectors.js
@@ -1,34 +1,16 @@
 // Local Dependencies
-import { documentCategories } from 'store/constants/reader';
-import { documentsView, formatTagOptions, formatTagValue, formatCategoryName } from 'utils/reader';
-import { createSelector } from 'reselect';
+import { documentCategories } from './constants/reader';
+import { documentsView, formatTagOptions, formatTagValue, formatCategoryName } from '../utils/index';
 import { isEmpty } from 'lodash';
 
 /**
  * Filtered Documents state
  */
-
-const getFilteredDocIds = (state) => state.reader.documentList.filteredDocIds;
-const getAllDocs = (state) => state.reader.documentList.documents;
-const getView = (state) => state.reader.documentList.view;
-const getSelectedDoc = (state) => state.reader.documentViewer.selected;
-const getFilterCriteria = (state) => state.reader.documentList.filterCriteria
-
-export const getFilteredDocuments = createSelector(
-  [getFilteredDocIds, getAllDocs],
-  (filteredDocIds, allDocs) => filteredDocIds.reduce(
-    (list, id) => ({ ...list, [id]: allDocs[id] }),
+export const filteredDocuments = ({ reader }) =>
+  reader.documentList.filteredDocIds.reduce(
+    (list, id) => ({ ...list, [id]: reader.documentList.documents[id] }),
     {}
-  )
-)
-
-export const getdocsFiltered = createSelector(
-  [getFilterCriteria], (filterCriteria) => {
-    return filterCriteria.searchQuery ||
-      !isEmpty(filterCriteria.category) ||
-      !isEmpty(filterCriteria.tag)
-  }
-)
+  );
 
 /**
  * Selector for the Documents
@@ -36,13 +18,12 @@ export const getdocsFiltered = createSelector(
  * @returns {Object} -- The Documents
  */
 export const documentState = (state) => {
-  console.log("Feature Flag Enabled - Bulk Upload")
   // Set the filtered documents
-  const documents = getFilteredDocuments(state);
+  const documents = filteredDocuments(state);
 
   // Calculate the number of documents
-  const docsCount = getFilteredDocIds(state) ?
-    getFilteredDocIds(state).length :
+  const docsCount = state.reader.documentList.filteredDocIds ?
+    state.reader.documentList.filteredDocIds.length :
     Object.values(documents).length;
 
   // Return the Filtered Documents and count
@@ -54,35 +35,33 @@ export const documentState = (state) => {
  * @param {Object} state -- The current Redux Store state
  * @returns {Object} -- The Documents List State
  */
-export const documentListScreenMemoized = (state) => {
+export const documentListScreen = (state) => {
   // Get the filtered documents and count
   const { documents, docsCount } = documentState(state);
 
-  Object.values(getAllDocs(state)).
+  Object.values(state.reader.documentList.documents).
     reduce((list, doc) => [...list, ...doc.tags], []).
     filter((tags, index, list) => list.findIndex((tag) => tag.text === tags.text) === index);
-
-  const getDocumentsView = createSelector([getAllDocs, getFilterCriteria, getView],
-    (docs, filterCriteria, view) => {
-      return documentsView(Object.values(docs), filterCriteria, view);
-    });
-
-  const docsFiltered = getdocsFiltered(state);
-
-  const documentView = getDocumentsView(state);
 
   return {
     documents,
     docsCount,
-    docsFiltered,
-    tagOptions: formatTagOptions(getAllDocs(state)),
-    currentDocument: getSelectedDoc(state),
-    storeDocuments: getAllDocs(state),
+    docsFiltered:
+      state.reader.documentList.filterCriteria.searchQuery ||
+      !isEmpty(state.reader.documentList.filterCriteria.category) ||
+      !isEmpty(state.reader.documentList.filterCriteria.tag),
+    tagOptions: formatTagOptions(state.reader.documentList.documents),
+    currentDocument: state.reader.documentViewer.selected,
+    storeDocuments: state.reader.documentList.documents,
     documentList: state.reader.documentList,
     comments: state.reader.annotationLayer.comments,
-    documentsView: documentView,
-    filterCriteria: getFilterCriteria(state),
-    filteredDocIds: getFilteredDocIds(state),
+    documentsView: documentsView(
+      Object.values(documents),
+      state.reader.documentList.filterCriteria,
+      state.reader.documentList.view
+    ),
+    filterCriteria: state.reader.documentList.filterCriteria,
+    filteredDocIds: state.reader.documentList.filteredDocIds,
     searchCategoryHighlights:
       state.reader.documentList.searchCategoryHighlights,
     manifestVbmsFetchedAt: state.reader.documentList.manifestVbmsFetchedAt,
@@ -99,57 +78,41 @@ export const documentListScreenMemoized = (state) => {
  * @param {Object} state -- The current Redux Store state
  * @returns {Object} -- The Document State
  */
-export const documentScreenMemoized = (state) => {
+export const documentScreen = (state) => {
   // Get the filtered documents and count
   const { documents, docsCount } = documentState(state);
+  const categories = Object.keys(documentCategories).reduce((list, key) => {
+    // Set the current Category
+    const cat = state.reader.documentViewer.selected[formatCategoryName(key)] ? key : '';
 
-  const getCategories = createSelector(getSelectedDoc, (selectedDoc) => {
-    return Object.keys(documentCategories).reduce((list, key) => {
-      // Set the current Category
-      const cat = selectedDoc[formatCategoryName(key)] ? key : '';
-      // Return the Categories Object
-      return {
-        ...list,
-        [cat]: true
-      };
-    }, {})
-  })
+    // Return the Categories Object
+    return {
+      ...list,
+      [cat]: true
+    };
+  }, {});
 
-  const categories = getCategories(state);
   // Filter the comments for the current document
-
-  const getAllComments = state => state.reader.annotationLayer.comments
-  const getComments = createSelector([getAllComments, getSelectedDoc],
-    (allComments, selectedDoc) =>
-    allComments.filter((comment) =>
-    comment.document_id === selectedDoc.id)
-    );
-
-  const comments = getComments(state);
-  // Get the tag options for the current document
-  const getTagOptions = createSelector([getAllDocs],
-    (allDocs) => {
-      return formatTagValue(
-        formatTagOptions(allDocs)
-      )
-    })
-  const tagOptions = getTagOptions(state)
-
-  const docsFiltered = getdocsFiltered(state);
-
+  const comments = state.reader.annotationLayer.comments.filter((comment) =>
+    comment.document_id === state.reader.documentViewer.selected.id);
 
   return {
     documents,
     docsCount,
     categories,
     comments,
-    docsFiltered,
+    docsFiltered:
+      state.reader.documentList.filterCriteria.searchQuery ||
+      !isEmpty(state.reader.documentList.filterCriteria.category) ||
+      !isEmpty(state.reader.documentList.filterCriteria.tag),
     currentPageIndex: state.reader.documentViewer.currentPageIndex,
     pendingTag: state.reader.documentViewer.pendingTag,
     editingTag: state.reader.documentViewer.editingTag,
     pendingCategory: state.reader.documentViewer.pendingCategory,
     documentTags: state.reader.documentViewer.tags,
-    tagOptions,
+    tagOptions: formatTagValue(
+      formatTagOptions(state.reader.documentList.documents)
+    ),
     viewport: state.reader.documentViewer.viewport,
     keyboardInfoOpen: state.reader.documentViewer.keyboardInfoOpen,
     pendingDeletion: state.reader.annotationLayer.pendingDeletion,
@@ -164,14 +127,14 @@ export const documentScreenMemoized = (state) => {
     windowingOverscan: state.reader.documentViewer.windowingOverscan,
     deleteCommentId: state.reader.documentViewer.deleteCommentId,
     shareCommentId: state.reader.documentViewer.shareCommentId,
-    filterCriteria: getFilterCriteria(state),
+    filterCriteria: state.reader.documentList.filterCriteria,
     openSections: state.reader.documentViewer.openedAccordionSections,
-    currentDocument: getSelectedDoc(state),
-    filteredDocIds: getFilteredDocIds(state),
+    currentDocument: state.reader.documentViewer.selected,
+    filteredDocIds: state.reader.documentList.filteredDocIds,
     appeal: state.reader.appeal.selected,
     searchCategoryHighlights:
       state.reader.documentList.searchCategoryHighlights,
-    storeDocuments: getAllDocs(state),
+    storeDocuments: state.reader.documentList.documents,
     annotationLayer: state.reader.annotationLayer,
     hidePdfSidebar: state.reader.documentViewer.hidePdfSidebar,
     hideSearchBar: state.reader.documentViewer.hideSearchBar,
